use sprout's videocodec driver

+++ drivers/misc/mediatek/videocodec/mt8127/common/val/src/val.c
@@ -1,3 +1,17 @@
+/*
+* Copyright (C) 2011-2014 MediaTek Inc.
+* 
+* This program is free software: you can redistribute it and/or modify it under the terms of the 
+* GNU General Public License version 2 as published by the Free Software Foundation.
+* 
+* This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
+* without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+* See the GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License along with this program.
+* If not, see <http://www.gnu.org/licenses/>.
+*/
+
 /**
  * @file
  *   val.c
@@ -175,7 +189,7 @@ VAL_RESULT_T eVideoWaitEvent(
         status = VAL_RESULT_NO_ERROR;
     }
     else {
-        MFV_LOGE("eVideoWaitEvent wake up by %d", (int)i4Ret);
+        MFV_LOGE("eVideoWaitEvent wake up by %ld", i4Ret);
         status = VAL_RESULT_NO_ERROR;
     }
     *((VAL_UINT8_T *)a_prParam->pvReserved) = VAL_FALSE;
